requirements.txt
# People on the Move - Dependencies

# Web Framework
flask>=2.3.0

# Database
sqlalchemy>=2.0.0

# HTTP & Parsing
requests>=2.31.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Date handling
python-dateutil>=2.8.0

# CLI
click>=8.1.0

# AI/API integrations
anthropic>=0.18.0
newsapi-python>=0.2.7

# Configuration
python-dotenv>=1.0.0

# Production server
gunicorn>=21.0.0

# PostgreSQL support (for production)
psycopg2-binary>=2.9.0
